Changed IDE from QtCreator to CLion (cmake), added Runner for auto copy executable files to working dir

+++ Global/grender.cpp
@@ -1,4 +1,4 @@
-#include <grender.h>
+#include <Global/grender.h>
 GRender::GRender():QObject(){
 
 }
@@ -26,10 +26,10 @@ void GRenderer::render(QPaintDevice* dev){
     glColor3f(1.0, 1.0, 1.0);
     glOrtho(-1.0, 1.0, -1.0, 1.0, -1.0, 1.0);
     glBegin(GL_POLYGON);
-    glVertex2f(-0.5, -0.5);
-    glVertex2f(-0.5, 0.5);
+	glVertex2f((GLfloat) -0.5, (GLfloat) -0.5);
+	glVertex2f((GLfloat) -0.5, 0.5);
     glVertex2f(0.5, 0.5);
-    glVertex2f(0.5, -0.5);
+	glVertex2f(0.5, (GLfloat) -0.5);
     glEnd();
     glFlush();
